Refactor botón reporte afiliados
