+++ src/components/common/details/PageDetailHeader.jsx
@@ -5,8 +5,6 @@ import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 import { useNavigate } from 'react-router-dom';
 import WarningDeleteDialog from '../forms/WarningDeleteDialog';
 import { detailHeaderConfig } from '../../../utils/detailHeaderConfig';
-import { getReporteAfiliadoById } from '../../../services/afiliado';
-import ArrowCircleDownIcon from '@mui/icons-material/ArrowCircleDown';
 
 export default function PageDetailHeader({
   type,
@@ -38,18 +36,6 @@ export default function PageDetailHeader({
     }
   };
 
-  const downloadReport = async () => {
-    const pdf = await getReporteAfiliadoById(id);
-    const url = window.URL.createObjectURL(pdf);
-    const link = document.createElement('a');
-    link.href = url;
-    link.setAttribute('download', `reporteAfiliado${id}.pdf`);
-    document.body.appendChild(link);
-    link.click();
-    link.parentNode.removeChild(link);
-    window.URL.revokeObjectURL(url);
-  };
-
   return (
     <Box sx={{ mb: 4 }}>
       <Grid container alignItems="center" justifyContent="space-between">
@@ -70,17 +56,6 @@ export default function PageDetailHeader({
 
         {onDelete && (
           <Grid size={{ xs: 12, md: 'auto' }}>
-            {type === 'afiliado' && (
-              <Button
-                variant="contained"
-                color="primary"
-                startIcon={<ArrowCircleDownIcon />}
-                sx={{ textTransform: 'none', mr: 1 }}
-                onClick={downloadReport}
-              >
-                Descargar reporte
-              </Button>
-            )}
             <Button
               variant="contained"
               color="error"
