+++ src/utils/downloadFile.js
@@ -1,0 +1,11 @@
+export async function downloadFile(serviceFn, filename) {
+  const pdf = await serviceFn();
+  const url = window.URL.createObjectURL(pdf);
+  const link = document.createElement('a');
+  link.href = url;
+  link.setAttribute('download', filename);
+  document.body.appendChild(link);
+  link.click();
+  link.parentNode.removeChild(link);
+  window.URL.revokeObjectURL(url);
+}
